Address clang compiler warnings
Clang (by default) objects to the omission of parentheses in
expressios such as a && b || c, and also to brace elision
in aggregate initialisation. The latter complicates initialisation
of e.g. std::array. GCC has not turned this warning on for
that reason, and there is an open issue for this for clang:
https://llvm.org/bugs/show_bug.cgi?id=21629

* Add parentheses to resolve logical parentheses warning
* Suppress missing-braces diagnostic in test_optional.cpp