Treat whitespace lines in SWC file parser.
Whitespace-only lines are now treated correctly in SWC file parser. This
automatically leads to a consistent and nice handling of old mac-style
eol; they are just another whitespace (not newline).

Refined the test cases a bit.

Added support for mapping swc record types to nest::mc::segmentKind.
This required the adjustment in some tests + the ball_and_stick file, since
the dendrite type of swc record is type '3'. Before, everything except
soma was added (hardcoded) as a dendrite. An swc_parse_error will be
thrown in case of an impossible mapping. Also, no check is performed for
inconsistencies of SWC record types along a branch; the type of the
branch is assumed to be the type of the first record (compartment) in
the branch.

Finally, adapted to the coding style.