Add basic debugging trace capability.
Defines a TRACE(...) macro that, when enabled at
compile time with -DWITH_TRACE, prints to std::cerr
the location in the source, a time stamp, and the
values of any variables given to the macro.

Uses a mutex to protect against simultaneous writes
to std::cerr when TBB is enabled; it is not
optimized for speed.