modcc generates net_receive for CUDA backend
The CUDA back end was using the default net_receive implementation,
which was a "nop". This is why event based simulations were not
validating on the GPU.

We use a CPrinter to generate the same net_receive block that is
used for the multicore backend. This is not efficient, because
each read/write requires a cudamemcpy between host and device
memory, however it allows us to pass all unit and validation tests.

A more efficient GPU-specific implementation is left for later
optimization work.