Allow explicit make of GPU mechanisms.
When NMC_WITH_CUDA is off, still allow the generation
of CUDA mechanisms with modcc by exposing the
`build_all_gpu_mods` target; this can be performed
independently of the presence (or otherwise) of a
CUDA development environment.

This eases development of GPU-related modcc tasks,
as preliminary work can be performed and checked on
machines without a CUDA environment.