remove generic event interface on `cell_group`
`cell_group` had a templated method `cell_group::enqueue_events()`, which
was templated on the type of container used to pass the set of events
to enqueue.

The time_type that permeates the code is taken from spike,
which is itself a specialzied type alias of `basic_spike`. This is not
an intuitive location to define the `time_type`.

* define the default time type in `common_types.hpp`: `using default_time_type = float`
* use this in the definition: `using spike = basic_spike<..., default_time_type>`
* rename `postsynaptic_spike_event` -> `basic_postsynaptic_spike_event`
* define `postsynaptic_spike_event = basic_postsynaptic_spike_event<default_time_type>`
* standardise the `cell_group::enqueue_events(std::vector<post_synaptic_event>` to
  explicitly use

This approach isn't fully generic, because it doesn't flexibly allow
the use of types with custom `time_types`. However this wasn't the case before,
where there were some entrt points into the model that implicitly enforced
the use of `float`.
Using generic spike types via the runtime polymorphic interface will require
some more elaborate type-erasure plumbing anyway, so I like this approach of
making the time type more explicitly global for now.