Brunel Network miniapp (#324)
Miniapp for simulating the Brunel network of LIF neurons. The network consists of 3 populations: excitatory, inhibitory and Poisson population. Each neuron from the network receives a fixed number (proportional to the size of the population) of incoming connections from each of the three populations.

The parameters that can be passed as command-line arguments are the following:

-n (--n_excitatory): total number of cells in the excitatory population.

-m (--n_inhibitory): total number of cells in the inhibitory population.

-e (--n_external): total number of cells in the Poisson (external) population.

-p (--in_degree_prop): the proportions of connections from each of the 3 populations that each neurons receives.

-w (--weight): weight of all excitatory connections.

-g (--rel_inh_w): relative strength of inhibitory synapses with respect to the excitatory ones.

-d (--delay): the delay of all connections.

-r (--rate): rate of Poisson cells [kHz]

-t (--tfinal): length of the simulation period [ms]

-s (--delta_t): simulation time step [ms] (this parameter is ignored)

-G (--group-size): number of cells per cell group

-f (--spike-file-output): save spikes to file [Bool]

-z (--profile-only-zero): only output profile information for rank 0

-v (--verbose): present more verbose information to stdout

For example, we could run the command like:
./brunel_miniapp.exe -n 400 -m 100 -e 400 -p 0.05 -w 1.2 -d 0.1 -g 1.0 -r 1 -t 100 -s 1 -G 10 -f

========================================================================

The output of the simulation can be visualised using the provided python script plot_spikes_brunel.py. This script expects 3 command line arguments:

-n (--n_exc): total number of cells in the excitatory population.

-m (--n_inh): total number of cells in the inhibitory population.

-e (--n_ext): total number of cells in the Poisson (external) population.

which should match the values of the miniapp simulation. For example, if the miniapp was run as in the previous example, then this script should be invoked as:

python3 plot_spikes_brunel.py -n 400 -m 100 -e 400

This script will then produce the following plot:
brunel_spikes.pdf
where only a half of Poisson spikes are plotted for the sake of clarity.