LIF neurons: CPU backend with Brunel Network miniapp (#441)
Two main contributions:

1) Implementation of LIF neuron model with no kernel and no external input (I_e=0)

The input current to each neuron is therefore just the sum of all the weights of incoming spikes. We integrate in jumps dt = min(t_final - t, t_event - t), since we know the exact solution of the differential equal describing the membrane potential.

2) Miniapp for simulating the Brunel network of LIF neurons. 

The network consists of 2 main populations: excitatory and inhibitory populations. Each neuron from the network receives a fixed number (proportional to the size of the population) of incoming connections from both of these groups. In addition to the input from excitatory and inhibitory populations, each neuron receives a fixed number of connections from the Poisson neurons producing the Poisson-like input that is integrated into the LIF cell group so that the communication of Poisson events is bypassed.