Avoid intermediate underflow in expm1 calc with ftz.
Intel compiler with default options does not guarantee
correct fp behaviour with subnormals; it presumably
sets the fp state to flush to zero.

Reordering a multiply and divide in the expm1 calculation
avoids a transient subnormal value that was causing the
routine to incorrectly return zero for very small, but
normal, arguments.