Extend ring (#611)
Extend the ring benchmark to have an optional number of synapses attached to each cell, instead of a fixed count of one synapse per cell.
This doesn't change the behavior of the model: only the first synapse is used for communication. The other synapses only effect is to
increase the per-cell computational overheads, to more effectively mimic real world performance.