CV policies with unit tests.
* Add new `cv_policy` class that wraps classes derived from
`cv_policy_base`, which in turn provide a facility for producing CV
boundaries on morphology trees from a `cable_cell`.
* Implement fixed-CVs-per-branch and maximum-CV-length policies, with
flags that allow for single-CV roots and for CVs that terminate at fork
points or which are centred on fork points.
* Expose locset `join` expressions.
* Add `empty()` test method on `morphology`.
* Expose branch length on `em_morphology` via member function.
* Account for the length of a spherical root branch in `em_morphology`.
* Test for validity but do not canonicalize locations when thingifying.
* Rename `root_sample_has_same_tag_as_child` to
`root_sample_tag_differs_from_children` to better reflect sense of
test.
* Add unit tests for CV policies, branch_length method.