Add codecov reports as part of CI (#1020)
This is a bit experimental. It builds arbor with `-fprofile-arcs -ftest-coverage` and creates a docker image with just the project binaries, lcov, gcov and generated `.gcno` files. This should still be reasonably small since there are no object files etc. When run on daint, `.gcna` files are produced (should be thread / process safe according to gcc's manual) inside of the container, and `lcov` is used to combine all those things into a single file with a random name in the mounted git project folder (hopefully this solves clashes with multiple nodes). After all tests are run, the combined reports are uploaded to codecov.io.