Implement 'stitch' morphology builder.
* Add (forward) ordered forest implementation, tests.
* Add `segment` region expression; to ease implementation, `msegment`
now also knows its own id.
* Add `stitch_builder` and `stitched_morphology`. A stitch corresponds
to a labelled, linearly-interpolated segment which can be attached at
any point along a parent stitch. A `stitched_morphology` takes a
`stitch_builder` object and constructs a segment tree and morphology,
and provides a dictionary of stitch labels to segment indices and
region expressions.
* Add `import` method for `label_dict`, so that the label dictionary
returned by `stitched_morphology` can be merged with an existing
dictionary.
* Add section on stitch builder etc. to cable cell docs.
* Update cable cell docs to remove out of date info and to provide some
context.
* Describe ordered forest datastructure and interface in a long comment
at the beginning of the header file.