Expose `time` to the mechanisms  (#1113)
The `vec_t_` time vector is already available to the mechanisms, but not exposed. It is indexed by `cell_index` instead of directly by the `node_index` (`vec_t[vec_ci_[node_index_[i]]]`). This kind of indexing was previously unavailble. It has not been added to the printers. 
This PR also includes some cleanup in the vectorized code printer. 

Address #1109