📊 Pin threads (#2094)
Allow thread and process affinity to request process and thread affinity
via `proc_allocation`.
Affinity improves resource utilisation and thus performance, see
experiments below, mainly
of CPU-only simulations. All other scenarios will not pay for it, so
this is a straight win.

This makes `hwloc` an optional dependency. If found, we offer affinity,
else we don't. It's
commonly deployed on HPC clusters, though, so we have good chances of
finding it.
Some care is taken to interact nicely with MPI and SLURM.

Benchmarks on JUWELS-CPU (48 cores x 2 SMT) with a small-ish network of
~100 simple
cells, (release, SIMD).

![thread-scaling-GCL-juwels](https://user-images.githubusercontent.com/24411438/223042477-2fc21ddf-4a06-4d23-a50f-161c469b6aff.png)