Python/pip strikes again: execute `pip install` in venvs (#2111)
A recent release of `pip` merged a PR addressing PEP 668. Distros of
Python may now mark `site-packages` directories as 'managed', which
means `pip` is not allowed to write to them. This means Arbors CMake is
now, due to assumptions about prefixes, is not able to correctly install
the Python package anymore: `pip` is a hard requirement for that. We
already document that `pip` is the preferred way, so that's OK, but
what's new is that `pip` doesn't seem to default to a user install
anymore, if the system dir is not writable (or 'managed'), ergo: install
does not work. Moreover, on at least Debian 12, there's even no longer
such a thing as user installs, so even the `--user` flag does not help.

`venv`s seem therefore the only place to install Python packages these
days, so let's.